requirements.txt
-i https://pypi.org/simple
certifi==2022.6.15
charset-normalizer==2.0.12
dropbox==11.32.0
idna==3.3
ply==3.11
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
stone==3.3.1
urllib3==1.26.9
